Ref: extending call signature of depictRequest() helper.
